\name{ICD10cod}
\alias{ICD10cod}
\title{Tibble of CODs}
\description{ Returns a tibble with  cause of death (COD) names, codes in SEER, and ICD10 codes.} 
\usage{ICD10cod()} 

\value{Tibble with COD names, SEER codes and ICD10 codes.}


\author{Tom Radivoyevitch (radivot@ccf.org)}
\seealso{\code{\link{SEERaBomb-package} }}

\examples{
library(SEERaBomb)
ICD10cod() # returns the tibble
}


\keyword{IO}
